// projects/web-app/src/types/auction-types.ts
export type AuctionWonDetail = {
  auctionId: string;
  auctionEndTime: Date;
  auctionStartTime: Date;
  currentHighestBid: number;
  winnerId: string;
  reviews: Pick<AuctionReview, "id" | "reviewerId" | "rating" | "comment">[];
  taskId: string;
  taskName: string;
  taskDetail: string | null;
  taskStatus: TaskStatus;
  taskDeliveryMethod: string | null;
  taskImageUrl: string | null;
  creator: {
    creatorUserId: string;
    creatorAppUserName: string | null;
    creatorUserImage: string | null;
  };
  reporters: {
    reporterUserId: string;
    reporterAppUserName: string | null;
    reporterUserImage: string | null;
  }[];
  executors: {
    executorUserId: string;
    executorAppUserName: string | null;
    executorUserImage: string | null;
  }[];
};